NC: Add DYTC Controls to status bar app
Supports PSC Mode for more power settings
Also adds support to save and restore from the plist file. and it auto applies the setting on first boot, after sleep and hibernation so you dont have to manually set it again. 

This fixes an issue in my T480 that limits the CPU temps to 70c on the stock config and i had to change the setting each time I boot or wake from sleep.